Refactor so we can swap out the runDB function
The longer term plan here is that we will take the functions that
compose the queries, make those swappable, and use this as a point to
swap for a mock implementation when it comes to testing time. This
should let us test that we can supply inputs via the API as usual, but
swap out the backend database for mock data.